+++ extensions/world-clock/src/utils/costants.ts
@@ -4,9 +4,10 @@ export const TIMEZONE_BASE_URL = "https://worldtimeapi.org/api/timezone";
 export const IP_BASE_URL = "http://worldtimeapi.org/api/ip/";
 export const IP_GEOLOCATION_API = "http://ip-api.com/json/";
 
-export enum LOCALSTORAGE_KEY {
+export enum localStorageKey {
   STAR_TIMEZONE = "StarTimeZone",
   TIMEZONE_CACHE = "TimezoneCache",
+  SHOW_DETAILS = "ShowDetails",
 }
 
 export const filterTag = [