+++ extensions/world-clock/CHANGELOG.md
@@ -1,5 +1,10 @@
 # World Clock Changelog
 
+## [Fix 24hour bug] - {PR_MERGE_DATE}
+
+- Fix 24hour bug
+- Add new action for Query World Time: Toggle Details
+
 ## [Add feature: adding aliases to time zones] - 2022-06-10
 
 - Add feature: you can set aliases for starred time zones now